pyproject.toml
[project]
name = "belgian-transport-pipeline"
version = "0.1.0"
description = "End-to-end Belgian public transport data pipeline — iRail + De Lijn → Kafka → Spark → dbt → Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-storage-file-datalake>=12.23.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "flake8>=7.2.0",
    "black>=25.1.0",
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
